refactor: page 컴포넌트에 features 컴포넌트 import로 수정
