+++ src/app/goal/new/goal/page.tsx
@@ -1,24 +1,7 @@
-'use client';
+import { GoalForm } from '@/features/goal/components';
 
-import { useFormContext } from 'react-hook-form';
-import { useRouter } from 'next/navigation';
-
-import type { GoalFormValues } from '@/features/goal/types';
-
-const CreateTargetPage = () => {
-  const router = useRouter();
-  const { register } = useFormContext<GoalFormValues>();
-
-  const handleClickNextButton = () => {
-    router.push('/goal/new/date');
-  };
-
-  return (
-    <>
-      <input {...register('title')} type="text" placeholder="한줄 목표" />
-      <button onClick={handleClickNextButton}>다음</button>
-    </>
-  );
+const CreateGoalPage = () => {
+  return <GoalForm />;
 };
 
-export default CreateTargetPage;
+export default CreateGoalPage;
