+++ src/app/goal/new/subgoal/page.tsx
@@ -1,24 +1,7 @@
-'use client';
+import { SubgoalForm } from '@/features/goal/components';
 
-import { useFormContext } from 'react-hook-form';
-import { useRouter } from 'next/navigation';
-
-import type { GoalFormValues } from '@/features/goal/types';
-
-const CreateSubtaskPage = () => {
-  const router = useRouter();
-  const { register } = useFormContext<GoalFormValues>();
-
-  const handleClickNextButton = () => {
-    router.push('/goal/new/more');
-  };
-
-  return (
-    <>
-      <input {...register('subgoals')} type="text" placeholder="목표를 이루기 위한 세부 항목" />
-      <button onClick={handleClickNextButton}>다음</button>
-    </>
-  );
+const CreateSubgoalPage = () => {
+  return <SubgoalForm />;
 };
 
-export default CreateSubtaskPage;
+export default CreateSubgoalPage;
